+++ src/net-unix.c
@@ -1,5 +1,5 @@
 /* Networking for UNIX
-   Copyright (C) 2014, 2016 Markus Uhlin. All rights reserved.
+   Copyright (C) 2014-2017 Markus Uhlin. All rights reserved.
 
    Redistribution and use in source and binary forms, with or without
    modification, are permitted provided that the following conditions are met:
@@ -29,19 +29,16 @@
 
 #include "common.h"
 
-#include <sys/socket.h>
 #include <sys/select.h>
+#include <sys/socket.h>
 
 #include <pthread.h>
 #include <string.h>
-#include <unistd.h>
 
 #include "assertAPI.h"
 #include "errHand.h"
-#include "irc.h"
 #include "libUtils.h"
 #include "network.h"
-#include "printtext.h"
 #include "strdup_printf.h"
 
 int g_socket = -1;
@@ -51,48 +48,8 @@ static pthread_t listenThread_id;
 static void *
 listenThread_fn(void *arg)
 {
-    struct network_recv_context ctx;
-    ssize_t       bytes_received;
-    const size_t  recvbuf_size   = 2048;
-    char         *recvbuf        = xcalloc(recvbuf_size, 1);
-    char         *message_concat = NULL;
-    enum message_concat_state state = CONCAT_BUFFER_IS_EMPTY;
-    struct printtext_context ptext_ctx;
-
     (void) arg;
-
-    ctx.sock     = g_socket;
-    ctx.flags    = 0;
-    ctx.sec      = 5;
-    ctx.microsec = 0;
-
-    irc_init();
-
-    do {
-	BZERO(recvbuf, recvbuf_size);
-	if ((bytes_received = net_recv(&ctx, recvbuf, recvbuf_size - 1)) == -1) {
-	    goto out;
-	} else if (bytes_received > 0) {
-	    irc_handle_interpret_events(recvbuf, &message_concat, &state);
-	} else {
-	    /*empty*/;
-	}
-    } while (g_on_air);
-
-  out:
-    ptext_ctx.window     = g_active_window;
-    ptext_ctx.spec_type  = TYPE_SPEC1;
-    ptext_ctx.include_ts = true;
-    if (g_on_air) {
-	printtext(&ptext_ctx, "Connection to IRC server lost");
-	g_on_air = false;
-    }
-    printtext(&ptext_ctx, "Disconnected");
-    net_ssl_close();
-    (void) close(g_socket);
-    irc_deinit();
-    free_not_null(recvbuf);
-    free_not_null(message_concat);
+    net_irc_listen();
     return (0);
 }