+++ src/events/privmsg.c
@@ -202,6 +202,8 @@ event_privmsg(struct irc_message_compo *compo)
 	    printtext(&ctx, "%s%c%s%s%c%s %s",
 		Theme("nick_s1"), c, COLOR4, nick, NORMAL, Theme("nick_s2"),
 		msg);
+	    if (ctx.window != g_active_window)
+		broadcast_window_activity(ctx.window);
 	} else {
 	    printtext(&ctx, "%s%c%s%s%c%s %s",
 		Theme("nick_s1"), c, COLOR2, nick, NORMAL, Theme("nick_s2"),