+++ src/commands/misc.h
@@ -15,5 +15,6 @@ void cmd_ilist   (const char *);
 void cmd_who     (const char *);
 void cmd_rules   (const char *);
 void cmd_cycle   (const char *);
+void cmd_version (const char *);
 
 #endif