+++ src/commands/misc.h
@@ -16,5 +16,6 @@ void cmd_who     (const char *);
 void cmd_rules   (const char *);
 void cmd_cycle   (const char *);
 void cmd_version (const char *);
+void cmd_time    (const char *);
 
 #endif