+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log #
 All notable changes to this project will be documented in this file.
 
+## [Unreleased] ##
+### Added ###
+- WIN32 specific: added a resource script
+
 ## [1.3] - 2017-01-14 ##
 ### Added ###
 - Event 435. Undocumented in the RFC. (Cannot change nickname while