+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log #
 All notable changes to this project will be documented in this file.
 
-## [Unreleased] ##
+## [1.4] - 2017-02-03 ##
 ### Added ###
 - WIN32 specific: added a resource script
 - Broadcast channel activity if our nickname is highlighted